refactor(memberbase): extract maskIfNeeded to break circular import
MemberCard imported maskIfNeeded from ./index while index imported
MemberCard, creating a circular module dependency. Move maskIfNeeded and
maskedFields into a dedicated leaf module imported by both.
